+++ tensorcircuit/backends/tensorflow_backend.py
@@ -7,6 +7,7 @@ from functools import reduce, partial
 from operator import mul
 from typing import Any, Callable, Optional, Sequence, Tuple, Union
 
+from scipy.sparse import coo_matrix
 import tensornetwork
 from tensornetwork.backends.tensorflow import tensorflow_backend
 
@@ -326,6 +327,11 @@ class TensorFlowBackend(tensorflow_backend.TensorFlowBackend):  # type: ignore
         )
 
     def numpy(self, a: Tensor) -> Tensor:
+        if self.is_sparse(a):
+            return coo_matrix(
+                (a.values, (a.indices[:, 0], a.indices[:, 1])), shape=a.get_shape()
+            )
+
         return a.numpy()  # only valid in eager mode
 
     def i(self, dtype: Any = None) -> Tensor: